maillogs location, user extensions, pcre include
detect location of maillogs when configuring /etc/mailer.conf

if vpopmail_qmail_ext is set in toaster-watcher.conf, modify chkuser_settings.h when installing qmail, so that user extensions work.

For whatever reason, simscan stopped detecting the location of the pcre headers. Added --with-pcre-include to simscan configure if enable-regex is selected.